CI Fixing and Documentation Update
OpenSSL was replaced with the ring crate for generating secure random
numbers. In addition, a new helper function was added for getting Unix
timestamps in milliseconds. This helper function works with Miri by
returning the test pattern instead of time, which is unavailable when
Miri is run with the default parameters.

Smart Device Mobile API documentation updated.

Some dependencies we plan to use later are now commented out so that the
dependency check passes.

The wrong action badge URL was fixed for the README.md.