Adding SecurityKey Struct
This commit replaces both AuthorizationKey and SharedKey with the
SecurityKey struct. In addition, a new KeyBytes type was introduced,
which is used to store the content of the SecurityKey object.

Some of the needed functionality, unit testing, and documentation were
added for the SecurityKey struct.

The std::num::ParseIntError was added to the error module. In addition,
a custom SecurityKeyWrongSize error was also added.

The config module was updated to used use the new SecurityKey struct.