Serialization with Serde
Custom Serialize and Deserialize were implemented for the SecurityKey so
that keys are written as hex strings for human-readable formats.

Both DeviceConfig and DeviceInfo are now serializable. Their name
property was put first to appear in config files when written.

The display message for the ErrorKind::SecurityKeyWrongSize was changed
to work better in both human-readable and binary content. This way, the
same error can be used from deserialize when the wrong amount of bytes
is given.

Serde implementations were added as dev dependencies as they are
currently only needed for unit testing. The serde_json is going to be
used with the service application later.