Adding the Server Binary
This commit introduces server binary which uses the mobile_api library.
The rocket and rocket_okapi crates were added for use with the server
binary. Endpoints have been added with placeholder implementations that
still need to be appropriately implemented.

The schemars crate was added to implement a custom JsonSchema trait for
the SecurityKey.

The dotenv crate and the .env file were added for setting up development
environment variables.

The plan document was updated with a few new ideas.