Managed State and Status Information
DeviceState and BusyGuard structures were added to the state module. The
former is used as managed state object for the server, and the latter
allows endpoints conveniently set a busy message that is cleared when
the operation is complete.

A new device_status module was added with structures for collecting the
system information.

Command endpoints were updated to use a custom responder with
information about HTTP codes and content. Furthermore, endpoint
implementations now use the server's state object to set busy messages.

The device status endpoint was updated to use a state object to get
system information and send it to the client in JSON format.

The server's main file was updated to load device information. A new
information file is created if it is missing, and the user is requested
to edit the file to change at least the product name. The server does
not start until the product name is changed from the "EDIT ME."

The `SIFIS_HOME_PATH` was changed to the "target/sifis-home" in the
`.env` file so that it is clear that it is not part of the source files.