Create Device Info App
This commit adds a new `create_device_info` binary to the project. This
binary can be used to create a `device.json` file, which is needed for
the `mobile_api_server` to run. In addition, the application can also
create a Qr Code of the authorization key as an SVG image.

The server program was updated to load `device.json` at the start and
tell about the `create_device_info`program if the file is missing.