Small Changes to Make CI Work with D-Bus
These changes should make CI testing work using the D-Bus for checking
that command scripts are run.

There is a known race condition issue with factory reset and
configuration endpoints. The problem is that both of these tests share
the same configuration file and are run parallel.

The integration test was intended to fix race condition problems, but it
started to turn out complex and unstable. Therefore, that path was
deemed unworthy. Instead, we plan to modify mobile_api to use paths
given to it instead of using the environment variables. The server
should still use the environment variable and tell it for the library.
This way, unit tests can run on separate folders, thus fixing race
condition problems.