Refactoring Mobile API Library
This commit contains changes to make the mobile_api library use an
instance object where paths are stored. This change allows unit tests to
use temporary folders to avoid race condition issues while testing.

Unit and integration tests were updated to match changes. In addition,
the integration test for create_device_info was improved to only a
write-protect folder to test error messages from the application. This
change allows the TempDir object to delete the temporary folder after
the test.
Finally, unit tests were moved back to the same file with
implementations, as this is the Rust coders' convention.

The helper application and server were updated to match changes to the
library.

More documentation was added.