Reorganization of the Project
These changes modify the project structure from workspaces to a single
project. This way, it is simpler to cross-compile and create Debian
packages for the Raspberry Pi.

Creating Debian packages was added for the project, which allows easy
setup for the target device. Package includes all needed binaries and
server files, but also systemd service.

The readme was updated to describe a new method for deploying a Mobile
API server and tools to a Raspberry Pi device.